# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ServicesHub B2B intermediation platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Admin operational and financial dashboard
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"
  /admin/requester-companies:
    get:
      operationId: listRequesterCompanies
      tags: [admin]
      summary: List all requester companies
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequesterCompany"
    post:
      operationId: createRequesterCompany
      tags: [admin]
      summary: Create a requester company
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequesterCompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequesterCompany"
  /admin/provider-companies:
    get:
      operationId: listProviderCompanies
      tags: [admin]
      summary: List all provider companies
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderCompany"
    post:
      operationId: createProviderCompany
      tags: [admin]
      summary: Create a provider company
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderCompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderCompany"
  /admin/work-orders:
    get:
      operationId: listAdminWorkOrders
      tags: [admin]
      summary: List all work orders (admin full view)
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: status
          in: query
          schema:
            type: string
        - name: requesterId
          in: query
          schema:
            type: string
        - name: providerId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrderAdminView"
  /admin/work-orders/{id}:
    get:
      operationId: getAdminWorkOrder
      tags: [admin]
      summary: Get full work order detail for admin
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderAdminView"
  /admin/work-orders/{id}/adjust:
    post:
      operationId: adminAdjustWorkOrder
      tags: [admin]
      summary: Admin adjust/reopen/cancel a work order
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAdjustWorkOrderInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderAdminView"
  /admin/work-orders/{id}/assign:
    post:
      operationId: adminAssignWorkOrder
      tags: [admin]
      summary: Admin assign a work order to a provider
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAssignWorkOrderInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderAdminView"
  /admin/invoices:
    get:
      operationId: listAdminInvoices
      tags: [admin]
      summary: List all invoices (admin)
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceAdminView"
  /admin/payments:
    get:
      operationId: listAdminPayments
      tags: [admin]
      summary: List all payments received from requesters
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /admin/payouts:
    get:
      operationId: listAdminPayouts
      tags: [admin]
      summary: List all payouts to providers
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payout"
    post:
      operationId: registerPayout
      tags: [admin]
      summary: Register a payout to a provider
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPayoutInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payout"
  /admin/travel-pricing:
    get:
      operationId: listTravelPricingRules
      tags: [admin]
      summary: List travel pricing rules
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TravelPricingRule"
    post:
      operationId: createTravelPricingRule
      tags: [admin]
      summary: Create a travel pricing rule
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTravelPricingRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelPricingRule"
  /admin/travel-pricing/{id}:
    put:
      operationId: updateTravelPricingRule
      tags: [admin]
      summary: Update a travel pricing rule
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTravelPricingRuleInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelPricingRule"
    delete:
      operationId: deleteTravelPricingRule
      tags: [admin]
      summary: Delete a travel pricing rule
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No content
  /admin/commission-settings:
    get:
      operationId: getCommissionSettings
      tags: [admin]
      summary: Get commission configuration
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionSettings"
    put:
      operationId: updateCommissionSettings
      tags: [admin]
      summary: Update commission settings
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionSettings"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionSettings"
  /admin/audit-log:
    get:
      operationId: listAuditLog
      tags: [admin]
      summary: List audit/history log for admin
      parameters:
        - $ref: "#/components/parameters/ProfileHeader"
        - name: workOrderId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"

  # ─── Notifications ────────────────────────────────────────────────
...
components:
  parameters:
    ProfileHeader:
      name: x-profile-id
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    RequesterCompany:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        taxId:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        cep:
          type: string
        city:
          type: string
        state:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, createdAt]
    CreateRequesterCompanyInput:
      type: object
      properties:
        name:
          type: string
        taxId:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        cep:
          type: string
        city:
          type: string
        state:
          type: string
      required: [name, email]
    ProviderCompany:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        taxId:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        cep:
          type: string
        city:
          type: string
        state:
          type: string
        commissionRate:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, createdAt]
    CreateProviderCompanyInput:
      type: object
      properties:
        name:
          type: string
        taxId:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        cep:
          type: string
        city:
          type: string
        state:
          type: string
        commissionRate:
          type: number
      required: [name, email]

    # ─── Catalog ─────────────────────────────────────────────────
...
    WorkOrderAdminView:
      type: object
      properties:
        id:
          type: string
        requesterCompanyId:
          type: string
        requesterCompanyName:
          type: string
        providerCompanyId:
          type: string
        providerCompanyName:
          type: string
        serviceCatalogItemId:
          type: string
        serviceName:
          type: string
        category:
          type: string
        location:
          type: string
        description:
          type: string
        notes:
          type: string
        status:
          $ref: "#/components/schemas/WorkOrderStatus"
        requestedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        basePrice:
          type: number
        travelCost:
          type: number
        commissionAmount:
          type: number
        finalPrice:
          type: number
        providerReceivable:
          type: number
        invoiceId:
          type: string
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistoryEntry"
      required: [id, serviceName, status, requestedAt]
...
    AdminAdjustWorkOrderInput:
      type: object
      properties:
        action:
          type: string
          enum: [reopen, cancel, correct]
        notes:
          type: string
        newStatus:
          $ref: "#/components/schemas/WorkOrderStatus"
      required: [action, notes]
    AdminAssignWorkOrderInput:
      type: object
      properties:
        providerCompanyId:
          type: string
        catalogItemId:
          type: string
      required: [providerCompanyId, catalogItemId]

    # ─── Financial ───────────────────────────────────────────────
...
    InvoiceAdminView:
      type: object
      properties:
        id:
          type: string
        workOrderId:
          type: string
        requesterCompanyName:
          type: string
        providerCompanyName:
          type: string
        serviceName:
          type: string
        basePrice:
          type: number
        travelCost:
          type: number
        commissionAmount:
          type: number
        finalPrice:
          type: number
        providerReceivable:
          type: number
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        generatedAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
      required: [id, workOrderId, finalPrice, status, generatedAt]
...
    Payment:
      type: object
      properties:
        id:
          type: string
        requesterCompanyId:
          type: string
        requesterCompanyName:
          type: string
        amount:
          type: number
        paymentMethod:
          type: string
        invoiceIds:
          type: array
          items:
            type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, requesterCompanyId, amount, paidAt, createdAt]
    RegisterPayoutInput:
      type: object
      properties:
        providerCompanyId:
          type: string
        workOrderIds:
          type: array
          items:
            type: string
        amount:
          type: number
        notes:
          type: string
      required: [providerCompanyId, workOrderIds, amount]
    Payout:
      type: object
      properties:
        id:
          type: string
        providerCompanyId:
          type: string
        providerCompanyName:
          type: string
        amount:
          type: number
        workOrderIds:
          type: array
          items:
            type: string
        notes:
          type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, providerCompanyId, amount, paidAt, createdAt]
...
    AdminDashboard:
      type: object
      properties:
        totalWorkOrders:
          type: integer
        openWorkOrders:
          type: integer
        completedWorkOrders:
          type: integer
        totalBilled:
          type: number
        totalReceived:
          type: number
        totalPayedOut:
          type: number
        totalCommission:
          type: number
        pendingPayments:
          type: number
        pendingPayouts:
          type: number
        recentWorkOrders:
          type: array
          items:
            $ref: "#/components/schemas/WorkOrderAdminView"
      required:
        - totalWorkOrders
        - openWorkOrders
        - completedWorkOrders
        - totalBilled
        - totalReceived
        - totalPayedOut
        - totalCommission
        - pendingPayments
        - pendingPayouts
...